Split the OpenGL layer out, pt 4: new Magnum::GL namespace.
At the moment just the GL library itself w/o the tests, and without
backwards compatibility aliases. The following types were left in the
root namespace, despite being in the GL/ directory, as they will get
moved back soon:

 * Image, CompressedImage and their dimensional typedefs
 * ImageView, CompressedImageView and their dimensional typedefs
 * PixelStorage

Not PixelFormat etc., that one will stay in the GL namespace and a
completely new PixelFormat enum will be provided in the root namespace.