Split the OpenGL layer out, pt 16: {Extensions::GL => GL::Extensions}.
Was Magnum::GL::Extensions::GL before and the redundancy was completely
unnecessary. Potential future extensions coming from GLX, EGL or whatnot
will most probably be in the Platform namespace in a completely separate
file, so this is not a problem.

All code internal to the GL library is affected, not much the outside,
as that is handled by the compatibility alias.